# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wikitruth API",
    "version": "1.1.0",
    "description": "Stable v1 contract for Wikitruth browser, civic-tenant, administrative, and accountable software-agent integrations."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Stable v1"
    },
    {
      "url": "/api",
      "description": "Compatibility alias"
    }
  ],
  "paths": {
...
    "/moderation/change-requests": {
      "get": {
        "summary": "List change requests for moderation",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit a suggestion-only change request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericMutationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/moderation/change-requests/{id}": {
      "put": {
        "summary": "Accept selected change-request fields or reject the request",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericMutationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiResponse"
                }
              }
            }
          },
          "409": {
            "description": "Stale base revision"
          }
        }
      }
    },
    "/moderation/rollback": {
      "post": {
        "summary": "Create a new revision by restoring a reviewed snapshot",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericMutationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rolled back",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/moderation/verdict-channel": {
      "put": {
        "summary": "Publish an explicit administrator final-say decision",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminFinalSayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verdict channel updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid channel, status, reasoning, or ethical framework"
          }
        },
        "description": "This exceptional route is unavailable to API clients and records an audited override of the current consensus snapshot."
      }
    },
    "/moderation/verdict-channels": {
      "put": {
        "summary": "Atomically update factual and ethical verdict channels",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericMutationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Both verdict channels updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Neither channel is updated when either payload is invalid"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
...
      "StandardApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "additionalProperties": true,
        "required": [
          "success"
        ]
      },
      "GenericMutationRequest": {
        "type": "object",
        "additionalProperties": true
      },
...
      "AdminFinalSayRequest": {
        "type": "object",
        "required": [
          "objectType",
          "objectId",
          "channel",
          "status",
          "reasoning",
          "overrideReason",
          "acknowledgeOverride"
        ],
        "properties": {
          "objectType": {
            "type": "integer"
          },
          "objectId": {
            "type": "string"
          },
          "channel": {
            "type": "string",
            "enum": [
              "factual",
              "ethical"
            ]
          },
          "status": {
            "type": "string"
          },
          "reasoning": {
            "type": "string",
            "minLength": 10
          },
          "framework": {
            "type": "string"
          },
          "evidenceRefs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "overrideReason": {
            "type": "string",
            "minLength": 10
          },
          "acknowledgeOverride": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AgentBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "wt_agent_<clientId>.<secret>",
        "description": "Scoped API-client credential tied to an accountable user"
      }
    },
...
  },
...
}